+++ scripts/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Added
 - [babel] Added rewrite-modules plugin for Babel 6
+- [gulp] Added strip-provides-module to strip `@providesModule` headers
 
 
 ## [0.5.0] - 2015-11-11