+++ scripts/CHANGELOG.md
@@ -4,6 +4,7 @@
 - [babel] Added rewrite-modules plugin for Babel 6
 - [gulp] Added strip-provides-module to strip `@providesModule` headers
 - [gulp] Added check-dependencies to ensure installed packages are compatible with package.json specification
+- [node] Added check-lib-requires script to ensure all lib files can be required
 
 
 ## [0.5.0] - 2015-11-11