+++ scripts/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Added
 - [babel] Added rewrite-modules plugin for Babel 6
+- [babel] Added dev-expression plugin for Babel 6
 - [gulp] Added strip-provides-module to strip `@providesModule` headers
 - [gulp] Added check-dependencies to ensure installed packages are compatible with package.json specification
 - [node] Added check-lib-requires script to ensure all lib files can be required